+++ module-api/src/test/java/com/family/ModuleApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.moduleapi;
+package com.family;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
