+++ module-stream/src/main/java/com/family/ModuleStreamApplication.java
@@ -1,4 +1,4 @@
-package com.modulestream;
+package com.family;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
