+++ module-stream/src/test/java/com/family/ModuleStreamApplicationTests.java
@@ -1,4 +1,4 @@
-package com.modulestream;
+package com.family;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
